Adjust and relocate messy files to better looking
