--- mobile/android/app/src/main/res/values/colors.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<resources>
-  <color name="splashscreen_background">#FFFFFF</color>
-</resources>
