--- mobile/android/react-settings-plugin/build.gradle.kts
@@ -1,19 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("jvm") version "1.9.24"
-    id("java-gradle-plugin")
-}
-
-repositories {
-    mavenCentral()
-}
-
-gradlePlugin {
-    plugins {
-        create("reactSettingsPlugin") {
-            id = "com.facebook.react.settings"
-            implementationClass = "expo.plugins.ReactSettingsPlugin"
-        }
-    }
-}
