--- mobile/android/react-settings-plugin/src/main/kotlin/expo/plugins/ReactSettingsPlugin.kt
@@ -1,10 +1,0 @@
-package expo.plugins
-
-import org.gradle.api.Plugin
-import org.gradle.api.initialization.Settings
-
-class ReactSettingsPlugin : Plugin<Settings> {
-  override fun apply(settings: Settings) {
-    // Do nothing, just register the plugin.
-  }
-}
