+++ mobile/styles/HomeScreenStyle.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  icon: {
+    width: 25,
+    height: 25,
+  },
+  tabBarLabel: {
+    fontSize: 12, // Adjust the size of the label text
+  },
+});
+
+export default styles;
