+++ mobile/styles/MessageLogStyle.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    padding: 10,
+  },
+  sectionHeader: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    backgroundColor: '#f4f4f4',
+    paddingVertical: 5,
+    paddingHorizontal: 10,
+  },
+});
+
+export default styles;
